Adds dynamic logos for namespaces
If there exists a page namespace:Logo, all pages in that namespace will use the first image on that page as their logo.

If there is no namespace:Logo page, they will use the first image on Mediawiki:logo.

If that doesn't exist, they'll fall back to the image defined in $wgLogo.

This still requires some optimisation, as there's currently a bit of FOUT